handle result of nextcloud_install script
It would be more useful to return result as boolean value in any case,
instead of the "success or exception" logic.

`nextcloud_install` in now executed on background (--daemon) because
the configure process takes quite a long time. And the request to
configure nextcloud from frontend usually timeout before we are able to
send result back.

Progress of configure action is reported back to frontend via WS,
by the script itself.

Nevertheless it is still good idea to check retval of the script
execution, because we can report result of the operation quickly.
Regardless whether it successfully started or failed.

The logic is following:
* retval == 0 --> script executed successfully and configuration is
  underway --> success
* retval != 0 --> something unexpected happened while executing script
  (before going to background) --> failure